requirements.txt
mariadb==1.1.10
sqlalchemy==2.0.32
cryptography==43.0.0
flask==3.0.3
flask-restx==1.3.0
numpy==2.0.1
opencv-python-headless==4.10.0.84
face-recognition-models @ git+https://github.com/ageitgey/face_recognition_models@e67de717267507d1e9246de95692eb8be736ab61
face-recognition==1.3.0
paho-mqtt==2.1.0
authlib==1.3.1
line-bot-sdk==3.11.1